+++ src/main/resources/db/migration/V1_01__create_companies_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS companies
 (
-    id       int                             NOT NULL AUTO_INCREMENT,
-    name     varchar(50)                     NOT NULL,
-    money    double(10, 2)                   NOT NULL,
+    id    int            NOT NULL AUTO_INCREMENT,
+    name  varchar(50)    NOT NULL,
+    money decimal(10, 2) NOT NULL,
     PRIMARY KEY (id)
 ) ENGINE = InnoDB
   DEFAULT CHARSET = utf8mb4
